refactor: update import paths for 'cn' and remove obsolete file
