+++ components/fumadocs/card.tsx
@@ -1,6 +1,6 @@
 import Link from 'fumadocs-core/link';
 import type { HTMLAttributes, ReactNode } from 'react';
-import { cn } from '../../lib/fumadocs/cn';
+import { cn } from '../../lib/cn';
 
 export function Cards(props: HTMLAttributes<HTMLDivElement>) {
   return (
