+++ components/fumadocs/files.tsx
@@ -3,7 +3,7 @@
 import { cva } from 'class-variance-authority';
 import { FileIcon, FolderIcon, FolderOpen } from 'lucide-react';
 import { type HTMLAttributes, type ReactNode, useState } from 'react';
-import { cn } from '../../lib/fumadocs/cn';
+import { cn } from '../../lib/cn';
 import {
   Collapsible,
   CollapsibleContent,
