--- lib/fumadocs/cn.ts
@@ -1,1 +1,0 @@
-export { twMerge as cn } from 'tailwind-merge';
